requirements.txt
requests~=2.32.4
openai~=1.61.1
python-dotenv~=0.21.0
jieba~=0.42.1
matplotlib~=3.9.2
wordcloud~=1.9.4
ruamel.yaml>=0.18.6